requirements.txt
Django==1.8.4
Fabric==1.10.2
Jinja2==2.8
Pillow==2.9.0
django-annoying==0.8.4
django-bower==5.0.4
django-countries==3.3
django-crispy-forms==1.5.2
django-taggit==0.17.1
django-wysiwyg-redactor==0.4.9
djrill==1.3.0
gunicorn==19.3.0
ipython==4.0.0
psycopg2==2.6.1
python-memcached==1.57
rollbar==0.9.14
sorl-thumbnail==12.3
